requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiopg==1.0.0
alembic==1.4.2
async-timeout==3.0.1
asyncpg==0.20.1
asyncpgsa==0.26.1
attrs==19.3.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.4.0
multidict==4.7.5
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.3
pytest-aiohttp==0.3.0
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.17
wcwidth==0.2.4
yarl==1.4.2